test: add some test to validate the behavior of guzzle
In order to check the behavior of the stack, I wanna check the behavior / capabilities
of the guzzle library. This also allow me to get a bit familiar with it. Those tests are
not correctly placed in the project and they probably fail since the target version of the
library is 7 and i cannot check the implementation on my machine which has a php 5.6 installed.
It remains to relocated test if they are relevant or use ideas in other implementations / tests.

Signed-off-by: ollie314 <mehdi.lefebvre@gmail.com>